Add files via upload
I selected the Titanic dataset. I have chosen to do an extensive analysis of the regression sector of the dataset. The multiple issues that I noticed were that there were inconsistencies in the fare and age columns. For example, in lines 181, 186, and 187, the ages of the guests were completely blank, and in the fares category, a couple of guests were shown to not pay a fee at all. However, the chart shows a value that insinuates that the guest was able to obtain a ticket without paying the fee for a ticket. This may seem farfetched, but considering the time in 1912 when the Titanic sank, it was certainly unconventional to board a cruise ship without paying a fee for the services that would be offered.

Upon typing the line of code to construct a regression model from my data, an error was shown displaying that I could no longer proceed as my table had blank data that would hinder the program from compiling the regression predictions upon the importation of the CSV file. In brief, the problems that the data set have were abnormalities based on logical circumstances, such as fares equating to a value of zero and a guest that appears to be without age according to the data.